Add files via upload
This Android Application was designed for the purpose of making the operations at the Doctor's site much more easier, organized and reliable. The application only provides access to the authorized Doctors and practitioners (currently using a sample data base for this task). After getting authenticated, the Doctor is directed to the next activity where he can scan a patient's Aadhaar QR code (Aadhaar UID used here for storing and maintaining data efficiently). On scanning, the presence of the patient's entry is checked in the database. If present, the patient's entire previous medical record is fetched from the firestore (cloud storage by google) and displayed for the doctor to view if required else a new entry is automatically added to the firestore database for the patient uniquely. The entries fetched are displayed by their date and time. Clicking on any one opens up a screen with all the details of that visit, the symptoms, medications given, tests taken. The doctor also has an option to add a new prescription which would get added to the previous patient data. There are also options for reminders which would be linked to the user's app login.  I am working on it along with the web scraping of the Doctor's authentic database and other modifications and features that would make the process far more improvised (for instance, a price comparison and an online-ordering module).